rename file: config -> serverConfig
